+++ examples/farViewer/CMakeLists.txt
@@ -60,7 +60,6 @@ _add_glfw_executable(farViewer
     gl_fontutils.cpp
     gl_mesh.cpp
     face_texture.cpp
-    hbr_refine.cpp
     "${SHADER_FILES}"
     "${INC_FILES}"
     $<TARGET_OBJECTS:sdc_obj>
